added diagnostic info, improved OSX install.

+++ install.sh
@@ -16,6 +16,8 @@
 ADB="/usr/bin/adb"
 FASTBOOT="/usr/bin/fastboot"
 UDEV="/etc/udev/rules.d/51-android.rules"
+OUTPUT=$(arch -1)
+OS=$(uname -1)
 
 # get sudo
 
@@ -96,10 +98,14 @@ elif [ "$(uname)" == "Darwin" ]; then # Mac OS X
     sudo curl -s -o $FASTBOOT "http://github.com/corbindavenport/nexus-tools/blob/master/bin/mac-fastboot?raw=true" -LOk
     echo "[INFO] Making ADB and Fastboot executable..."
     echo "[INFO] Downloading udev list..."
-    sudo curl -s -o $UDEV "http://github.com/corbindavenport/nexus-tools/blob/master/udev.txt" -LOk
-    sudo chmod 644 $UDEV
-    sudo chown root. $UDEV
-    sudo killall adb
+    if [ -n "$UDEV" ]; then
+        sudo curl -s -o $UDEV "http://github.com/corbindavenport/nexus-tools/blob/master/udev.txt" -LOk
+        sudo chmod 644   $UDEV
+        sudo chown root. $UDEV
+        sudo service udev restart
+        sudo killall adb
+    fi
+    echo "[INFO] Making ADB and Fastboot executable..."
     sudo chmod +x $ADB
     sudo chmod +x $FASTBOOT
     echo "[ OK ] Done!"
@@ -139,6 +145,8 @@ elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then # Generic Linux
     echo " "
     exit 0
 else
-    echo "[EROR] Your operating system could not be detected. Now exiting."
+    echo "[EROR] Your operating system or architecture could not be detected."
+    echo "[ERROR] OS: $OS"
+    echo "[EROR] ARCH: $ARCH"
     exit 1
 fi
